+++ app.js
@@ -49,109 +49,7 @@ router.route('/defact/drawing/').get(function (req, res) {
     });
 });
 
-//하자 리스트 이동 라우터
-router.route('/defact/list/').get(function (req, res) {
-    var selected_dong = req.query.dong,
-        selected_ho = req.query.ho,
-        selected_loc = req.query.loc;
-    var selectDefactSql = "select * from defact where dong=? and ho = ? and room=?;";
 
-    mySqlClient.query(selectDefactSql, [selected_dong, selected_ho, selected_loc], function (err, rows, fields) {
-        if (err) {
-            console.log("ERROR>>" + err);
-        } else {
-            var unsolvedDefact = [];
-            var solvedDefact = [];
-
-            rows.forEach(function (element) {
-
-                if (element.is_solved == '0') {
-                    unsolvedDefact.push(element);
-                } else {
-                    solvedDefact.push(element);
-                }
-
-            });
-            fs.readFile('./public/list_defact.html', 'utf8', function (error, data) {
-
-                res.send(ejs.render(data, {
-                    dong: selected_dong,
-                    ho: selected_ho,
-                    loc: selected_loc,
-                    unsolvedDefact: unsolvedDefact,
-                    solvedDefact: solvedDefact
-                }));
-
-            });
-        }
-    });
-});
-
-//하자 세부 페이지 라우터
-router.route('/defact/detail/').get(function (req, res) {
-
-    var defactId = req.query.id;
-
-    var selectDetailSql = 'select d.id, d.img, d.construction_name, d.construction_type, d.info, d.due_date, d.is_read, d.is_solved from defact d where d.id = ?';
-
-    var selectCommentSql = 'select u.name, u.type user_type,  c.comment from comment c, user u where c.defact_id = ? and u.id = c.user_id;';
-    mySqlClient.query(selectDetailSql, defactId, function (err, row) {
-        if (err) {
-            console.log("select detail sql ERROR>> " + err);
-        } else {
-            var detailInfo = row[0];
-            var commentInfo = [];
-            mySqlClient.query(selectCommentSql, defactId, function (err, rows) {
-                if (err) {
-                    console.log('select comment sql error>' + err);
-                } else {
-                    rows.forEach(function (element) {
-                        commentInfo.push(element);
-                    });
-                    fs.readFile('./public/detail_defact.html', 'utf8', function (err, data) {
-
-                        res.send(ejs.render(data, {
-                            detailInfo: detailInfo,
-                            commentInfo: commentInfo
-                        }));
-                    });
-                }
-
-            });
-        }
-    });
-});
-
-//하자 상세 페이지 댓글 등록 라우터
-router.route('/defact/add/comment/').post(function (req, res) {
-    if (req.session.user) {
-        var user = {
-            userId: req.session.user.id,
-            userType: req.session.user.userType,
-            userName: req.session.user.userName
-        };
-        var defactId = req.body.defact_id;
-        var inputComment = req.body.inputComment;
-        var insertCommentSql = 'insert into comment  (user_id, defact_id, comment )values (?,?,?)';
-
-        var params = [user.userId, defactId, inputComment];
-        mySqlClient.query(insertCommentSql, params, function (err) {
-            if (err) {
-                console.log("Comment Sql Error>>" + err);
-            } else {
-                //나중에 Ajax로 바꿔볼것!!!!!!
-                res.writeHead(302, {
-                    'Location': '/defact/detail?id=' + defactId
-                });
-                res.end();
-
-
-            }
-        });
-    } else {
-        res.send('<script type="text/javascript">alert("로그인 후 이용하세요."); window.location="/";</script>');
-    }
-});
 
 //하자 등록 이동 라우터
 router.route('/defact/add/').get(function (req, res) {
@@ -169,38 +67,6 @@ router.route('/defact/add/').get(function (req, res) {
     });
 });
 
-//--------------------------로그인-------------------------------------
-router.route('/process/login').post();
-
-//공사종류, 동, 호  선택페이지
-router.route('/select').get(function (req, res) {
-    if (req.session.user) {
-        //동, 호 선택
-        if (req.query.ctype) {
-            fs.readFile('./public/select.html','utf8',function(err,data){
-                //선택한 공사종류를 쿠키에 저장
-               res.cookie('ctypeCookie',req.query.ctype);
-                res.send(ejs.render(data, {
-                    name: req.session.user.userName,
-                    type: req.session.user.userType
-                }));
-            });
-        } 
-        //공사종류 선택
-        else {
-            fs.readFile('./public/select_const.html', 'utf8', function (error, data) {
-                var ctype = req.query.ctype;
-                res.send(ejs.render(data, {
-                    name: req.session.user.userName,
-                    type: req.session.user.userType
-                }));
-            });
-        }
-
-    } else {
-        res.send('<script type="text/javascript">alert("로그인 후 이용하세요."); window.location="/";</script>');
-    }
-});
 
 //하자 등록 라우터
 router.route('/defact/add_submit').post(function (req, res) {
@@ -219,16 +85,28 @@ router.route('/defact/add_submit').post(function (req, res) {
         tel: tel,
         type: type
     };
-
 });
 
 var index = require('./routes/index.js');
 router.route('/').get(index);
+
 var register = require('./routes/register.js');
 router.route('/register').get(register.register);
 router.route('/reg_submit').post(register.reg_submit);
 var login = require('./routes/login.js');
 router.route('/process/login').post(login);
+
+//공사종류, 동, 호  선택페이지
+var select = require('./routes/select.js');
+router.route('/select').get(select);
+
+//하자 리스트 이동 라우터
+var defact = require('./routes/defact.js');
+router.route('/defact/list/').get(defact.defactList);
+//하자 세부 페이지 라우터
+router.route('/defact/detail/').get(defact.defactDetailList);
+//하자 상세 페이지 댓글 등록 라우터
+router.route('/defact/add/comment/').post(defact.defactAddComment);
 
 app.use('/', router);
